+++ packages/blip/app/config.js
@@ -41,7 +41,8 @@ const defaultConfig = {
   YLPZ_RA_LAD_001_NL_REV: '0',
   YLPZ_RA_LAD_001_IT_REV: '0',
   YLPZ_RA_LAD_001_DE_REV: '0',
-  YLPZ_RA_LAD_001_ES_REV: '0'
+  YLPZ_RA_LAD_001_ES_REV: '0',
+  IDLE_TIMEOUT_MS: 1800000
 }
 
 /** @typedef {typeof defaultConfig} AppConfig */