+++ packages/yourloops/lib/config/config.ts
@@ -63,7 +63,8 @@ const defaultConfig: AppConfig = {
   YLPZ_RA_LAD_001_NL_REV: '0',
   YLPZ_RA_LAD_001_IT_REV: '0',
   YLPZ_RA_LAD_001_ES_REV: '0',
-  YLPZ_RA_LAD_001_DE_REV: '0'
+  YLPZ_RA_LAD_001_DE_REV: '0',
+  IDLE_TIMEOUT_MS: 1800000
 }
 const appConfig = _.assign({}, defaultConfig)
 if (_.has(window, 'config') && _.isObjectLike(_.get(window, 'config', null))) {