+++ packages/yourloops/test/integration/mock/auth0.hook.mock.ts
@@ -32,6 +32,7 @@ import { AuthenticatedUserMetadata } from '../../../lib/auth/models/enums/authen
 export const loggedInUserId = '919b1575bad58'
 export const loggedInUserEmail = 'john.doe@example.com'
 export const getAccessTokenWithPopupMock = jest.fn()
+export const logoutMock = jest.fn()
 
 export const mockAuth0Hook = (role: UserRoles = UserRoles.hcp, userId = loggedInUserId) => {
   (auth0Mock.useAuth0 as jest.Mock).mockReturnValue({
@@ -43,6 +44,7 @@ export const mockAuth0Hook = (role: UserRoles = UserRoles.hcp, userId = loggedIn
       sub: 'auth0|' + userId,
       [AuthenticatedUserMetadata.Roles]: [role]
     },
-    getAccessTokenWithPopup: getAccessTokenWithPopupMock
+    getAccessTokenWithPopup: getAccessTokenWithPopupMock,
+    logout: logoutMock
   })
 }